2.dataproc-spark-processing/requirements.txt
google-cloud-storage==2.5.0
google-cloud-dataproc==2.5.0
google-cloud-bigquery==2.34.0
google-cloud-biglake==0.1.0
pandas==1.3.5
pyyaml==6.0
kaggle==1.6.17